+++ src/library/Attempt/exports.ts
@@ -32,4 +32,4 @@ export {
   Attempt_thatEventually as thatEventually,
   type Attempt_EndGetter as EndGetter,
   type Attempt_EndRetriever as EndRetriever,
-} from './factory';
+} from './factory/exports';
